+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## Unreleased
 
+## 6.2.0.0
+
+- WindowsEventLog:
+  - Migrated the xWinEventLog from [xWinEventLog](https://github.com/PowerShell/xWinEventLog)
+    and renamed to WindowsEventLog.
+  - Moved strings in localization file.
+  - LogMode is now set with Limit-EventLog,
+  - Fixes [Issue #18](https://github.com/PowerShell/ComputerManagementDsc/issues/18).
+
 ## 6.1.0.0
 
 - Updated LICENSE file to match the Microsoft Open Source Team standard.